Move the 'TerminationHandler' class to a dedicated script file
